separated folder front / server
