+++ src/frontend/assets/donations.json
