+++ src/frontend/utils/trpc.ts
@@ -1,4 +1,4 @@
 import { createTRPCReact } from "@trpc/react-query";
-import type { AppRouter } from "../server";
+import type { AppRouter } from "../../server";
 
 export const trpc = createTRPCReact<AppRouter>();
